requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
duckdb>=0.10.0
pyarrow>=14.0.0
requests>=2.31.0
lxml>=5.0.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
fredapi>=0.5.0
